+++ ICEngine/src/Graphics/Gui_private.h
@@ -3,7 +3,7 @@
 
 #include "Texture.h"
 
-void ICE_Gui_UpdateTexture(int man, int gui);
+void ICE_Gui_UpdateTexture(size_t man, size_t gui);
 void ICE_Gui_Rect(ICE_Texture * texture, struct ICE_Box rect);
 
 #endif
